cleitan-os-1/requirements.txt
requests
beautifulsoup4
google-genai
selenium
python-dotenv
rich
Pillow
pygame